Regional

+++ hook/delete.php
@@ -1,12 +1,18 @@
 <?php
+
+use Mixin\Hell;
+
 require_once '../tools/warframe.php';
 $session->is_auth();
 
 if ( isset($_GET['model']) ) {
-    $form = new $_GET['model'];
-    unset($_GET['model']);
-    $form->delete($_GET['id']);
-}else {
-    Mixin\error('403');
-}
+
+    if (class_exists($_GET['model'])) {
+        $form = new $_GET['model'];
+        unset($_GET['model']);
+        $form->delete($_GET['id']);
+    }else dd("Модель не найдена!");
+
+}else Hell::error('403');
+
 ?>
